requirements.txt
clipspy==1.0.6
flask==3.1.2